s3_cdk_template/requirements.txt
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.cloudformation_include
